[K/N] Make initWithExternalRCRef to search for best-fitting class.
`-[KotlinBase initWithExternalRCRef:ref]` will search for the
best-fitting `Class` for the given `ref`, and construct the `KotlinBase`
instance with that `Class` instead of the current `[self class]`.

The code panics if the best-fitting `Class` turns out not descend
from `[self class]`: this happens when there's a class not generated
by Swift Export inheriting from an exported open class. This inheritance
is currently unsupported.

^KT-70680