[FIR] Use declaration-site session for performFirstRoundOfBoundsResolution
Consider a Java class A in library module `lib` that is referenced by
a module `app`.
When A is enhanced in a scope from module `app`, we mutate A's type
parameter bounds.
The session that is used in the call to resolveIfJavaType, can be
captured by the lazy type annotation argument mapping.
This is an issue in scenarios like Find Usages where many sessions
are created and analyzed.
Leaking use-site sessions that could otherwise be collected creates
memory pressure and performance problems.

#KT-70965 Fixed