IR Inliner, Native: Fix caching of inline functions before inlining
We need to make sure that every inline function is cached immediately
before being inlined in the state that was immediately before inlining.

Previously, such caching worked well only for private inline functions.
It was possible to gain a situation when a non-private inline function
is not cached despite being already lowered and read-to-go for
the inliner.

Now, with two new explicit Native phases "cacheOnlyPrivateFunctionsPhase"
and "cacheAllFunctionsPhase" all inline functions are guaranteed
to be cached.

#KT-70583 Fixed
#KT-67220