[CFG] Include Array constructor when calculating invocation kind
THe Array constructor is special and considered inline. This requires
special consideration when calculating inline status of calls to the
constructor. This helps CFG understand that variables are not captured
by value within the lambda passed to the constructor, but rather the
lambdas are inlined which delays variable initialization requirements.

^KT-70724 Fixed