[FIR] Fix getting fir for name in qualified typed reference
'qual.A<Ty>' has exactly one fir element (FirResolvedQualifier).
When trying to get the fir element for KtElement 'A',
we would previously return the mapped element for
'A<Ty>' which is a KtCallExpression, but which has no fir element.
This commit makes it so that the parent (KtQualifiedExpression) is used
if the call has no argument-list.

^KT-69819 Fixed