K2: match FunctionalTypeUtils function receivers with require contracts
Before this commit, we had several functions in FunctionalTypeUtils.kt
with ConeKotlinType receiver, that in fact required a ConeClassLikeType
receiver, as they had such a 'require' call at the first line.
In some corner cases, this could lead to an incorrect receiver type
and a runtime error in compiler.

This commit matches receiver types with real requirements
and fixes use-site receiver types, when it's necessary,
thus removing a possibility of such exceptions.

#KT-71228 Fixed