[KLIB Resolver] Change the order inside the resolution sequence
First, try to resolve a library using search roots by treating
`givenName` as the relative path. Then, only as a last resort,
try to resolve the library treating `givenName` as a unique name.

The motivation here is to give the preference to searching through
search roots over searching by unique name among user-specified
libraries, as the latter we are going to drop in the future.

^KT-67448