[PowerAssert] Read source text via PSI or FIR instead of the file
By reading the text from PSI or FIR, this removes a giant hack needed to
access the source code for tests. This should allow implementing
multi-module tests as well! Also, this solves a long-standing hack for
line endings which include carriage returns since we are reading the
same text that all the offsets are calculated from.

^KT-41888 Fixed