[PowerAssert] Use PersistentList to reduce list copying
When rewriting the IR with all the temporary variables, there is a lot
of list copying happening due to all the potential branching. By using
PersistentList, we should be able to reuse list segments whenever
possible and avoid the extra memory use. Probably not significant, but
still worth doing, just in case.