JVM: remove ReplaceDefaultImplsOverriddenSymbols
This lowering phase did not do anything useful, which is mostly proven
by the fact that removing it did not break anything.

In fact it never worked as expected because it was not applied to
classes from dependencies (because external IR is not lowered).
Supposedly the subsequent code (mostly bridge lowering) has learnt to
deal with the incorrect structure of fake overrides anyway, so for
consistency it's better to remove it.